apply DTO and AutoMapper

+++ StudentManagement/Services/StudentService/StudentService.cs
@@ -1,44 +1,55 @@
 ﻿
 
+using AutoMapper;
 using Microsoft.EntityFrameworkCore;
 using StudentManagement.Data;
+using StudentManagement.DTO;
 
 namespace StudentManagement.Services.StudentService
 {
     public class StudentService : IStudentService 
     {
-        private static List<Student> students = new List<Student> {
-             new Student(),
-             new Student { Id =1, Name= "John"},
-        };
         private readonly DataContext _context;
+        private readonly IMapper _mapper;
 
-
-        public StudentService(DataContext context)
+        public StudentService(DataContext context, IMapper mapper)
         {
             _context = context;
+            _mapper = mapper;
         }
 
-        public async Task<Response<List<Student>>> AddStudent(Student newStudent)
+        public async Task<Response<List<GetStudentDTO>>> AddStudent(AddStudentDTO newStudent)
         {  
-          var response= new Response<List<Student>>();
-          _context.Students.Add(newStudent);
-          await _context.SaveChangesAsync();
-          response.Data = await _context.Students.ToListAsync();
-          return response;
+            var response= new Response<List<GetStudentDTO>>();
+            var student = _mapper.Map<Student>(newStudent);          
+            _context.Students.Add(student);
+
+            await _context.SaveChangesAsync();
+            response.Data = await _context
+                .Students
+                .Select(student => _mapper.Map<GetStudentDTO>(student))
+                .ToListAsync();
+            return response;
         }
 
-        public async Task<Response<List<Student>>> DeleteStudent(int id)
+        public async Task<Response<List<GetStudentDTO>>> DeleteStudent(int id)
         {
-            var response = new Response<List<Student>>();
+            var response = new Response<List<GetStudentDTO>>();
 
             try
             {
-                var student = await _context.Students.FirstAsync(c => c.Id == id);
+                var student = await _context.Students.FirstOrDefaultAsync(c => c.Id == id);
+                if(student == null){
+                    throw new Exception("Student is not exist");
+                }
                 _context.Students.Remove(student);
                 await _context.SaveChangesAsync();
 
-                response.Data = await _context.Students.ToListAsync();
+                var students = await _context
+                    .Students
+                    .Select(student => _mapper.Map<GetStudentDTO>(student))
+                    .ToListAsync();
+                response.Data = students;
             }
             catch(Exception ex)
             {
@@ -49,46 +60,74 @@ namespace StudentManagement.Services.StudentService
             return response;
         }
 
-        public async Task<Response<List<Student>>> GetAllStudents()
+        public async Task<Response<List<GetStudentDTO>>> GetAllStudents()
         {
-            var dbStudents = await _context.Students.ToListAsync();
-            var response = new Response<List<Student>>();
+            var response = new Response<List<GetStudentDTO>>();
+            try
+            {
+                var dbStudents = await _context
+                    .Students
+                    .Select(student => _mapper.Map<GetStudentDTO>(student))
+                    .ToListAsync();
 
-            response.Data = dbStudents;
+                response.Data = dbStudents;
+            }
+
+            catch(Exception ex)
+            {
+                response.Success = false;
+                response.Message = ex.Message; 
+            }
+
+            
             return response;
         }
 
-        public async Task<Response<Student>> GetStudentById(int id)
+        public async Task<Response<GetStudentDTO>> GetStudentById(int id)
         {
-            var dbStudent = await _context.Students.FirstOrDefaultAsync(c => c.Id == id);
-            var response = new Response<Student>();
-      
-            response.Data = dbStudent;
+            var response = new Response<GetStudentDTO>();
+            try { 
+                var dbStudent = await _context.Students.FirstOrDefaultAsync(c => c.Id == id);
+                if(dbStudent == null)
+                {
+                    throw new Exception("Student is not exist");
+                }
+                response.Data = _mapper.Map<GetStudentDTO>(dbStudent);
+            }     
+            catch(Exception ex)
+            {
+                response.Success = false;
+                response.Message = ex.Message;
+            }
+
+
             return response;
         }
 
         
-         public async Task<Response<Student>> UpdateStudent(Student updatedStudent)
+         public async Task<Response<GetStudentDTO>> UpdateStudent(UpdateStudentDTO updatedStudent)
          {
-            var response = new Response<Student>();
+            var response = new Response<GetStudentDTO>();
 
-            try { 
-            var student = await _context.Students.FirstOrDefaultAsync(c => c.Id == updatedStudent.Id);
-            student.Name = updatedStudent.Name;
-            student.Gender = updatedStudent.Gender;
-            student.Course_Id = updatedStudent.Course_Id;
-
-            await _context.SaveChangesAsync();
-            response.Data = student;
+            try {
+                var student = await _context.Students.FirstOrDefaultAsync(c => c.Id == updatedStudent.Id);
+                if (student == null)
+                {
+                    throw new Exception("Student is not exist");
+                }
+                _mapper.Map(updatedStudent, student);
+                 await _context.SaveChangesAsync();
+                 response.Data = _mapper.Map<GetStudentDTO>(student);                         
             }
             catch(Exception ex)
+
             {
                 response.Message = ex.Message;
                 response.Success = false; 
             }
 
             return response;
-        }
+         }
         
     }
 }
